util: enhance dump tool
This commit enhances the efficiency of dump tool. Now it could process
several gigabytes of data within a minute. The JSON data is directly
piped into stdout instead of being cached in a stringstream. And the
analysis script is now using high-performance JSON parser while showing
progress when processing data.

Signed-off-by: Alex Chi <iskyzh@gmail.com>